v2 refactor-metadata field, file check, code cleanup,..
